+++ src/components/Header.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const Header = () => {
+  return (
+    <header className="bg-gray-800 text-white p-4 text-center">
+      <h1 className="text-3xl font-bold">Boids Simulation</h1>
+      <p className="text-lg">A simulation of flocking behavior using p5.js</p>
+    </header>
+  );
+};
+
+export default Header;
